Simplified mqtt java side implementation

+++ iot-with-ESP-MQTT/src/main/java/utils/TemperatureSample.java
@@ -4,9 +4,9 @@ import java.util.Date;
 
 public class TemperatureSample {
 	private final Date date;
-	private final int temperature;
+	private final double temperature;
 	
-	public TemperatureSample(Date date, int temperature) {
+	public TemperatureSample(Date date, double temperature) {
 		this.temperature = temperature;
 		this.date = date;
 	}
